requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.0
numpy==1.23.5
torch==2.0.1 --extra-index-url https://download.pytorch.org/whl/cpu
opencv-python-headless==4.8.0.76
deepface==0.0.75
tensorflow-cpu==2.8.0  # Використовуємо CPU версію TensorFlow
protobuf==3.20.0
requests==2.31.0
pydantic==2.3.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
gdown==4.7.1